% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burst_freq.R
\name{burst_freq}
\alias{burst_freq}
\title{Get burst frequencies of attacking individuals}
\usage{
burst_freq(df)
}
\arguments{
\item{df}{a dataframe with one variable named actor or winner}
}
\value{
a list with the burst lengths (number of consecutive attacks) by
    each individual
}
\description{
Get burst frequencies of attacking individuals
}
\examples{
burst_freq(chickens)
}
