% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chickens.R
\docType{data}
\name{chickens}
\alias{chickens}
\title{Example dataset of chicken behavior}
\format{A data frame with 271 observations and 4 variables:
\describe{
  \item{time}{Time in seconds of behavior}
  \item{behavior}{Behavior: P = peck; J = jump}
  \item{actor}{Actor}
  \item{recipient}{Recipient}
}}
\usage{
chickens
}
\description{
Dataframe with dominance behaviors directed from an actor
towards a recipient and timestampped.
}
\keyword{datasets}
