% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csf.R
\name{csf}
\alias{csf}
\title{Determine class-stability factor (CSF)}
\usage{
csf(df)
}
\arguments{
\item{df}{either a two-column dataframe of winners and losers or
a dataframe produced by `addstates` with columns winner, loser,
id and class}
}
\value{
a named vector of state ids with the CSF for all link numbers
}
\description{
Determine class-stability factor (CSF)
}
\examples{
df<-data.frame(winner=c(1,2,3,1,2,3,2,1,2,3,3,1,2,3,4,3,1,3,2,1,1,1,1,2,2),
loser=c(2,4,4,4,3,2,1,3,3,4,4,2,3,2,3,4,2,4,3,3,3,2,2,4,3)
)
csf(df)

csf(addstates(wldf()))
}
