% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csf_n.R
\name{csf_n}
\alias{csf_n}
\title{Determine class-stability factor (CSF) for specific number of links}
\usage{
csf_n(df, links = 4)
}
\arguments{
\item{df}{a dataframe produced by `addstates` with columns winner, loser,
id and class}

\item{links}{the number of links to determine the CSF for}
}
\value{
a named vector of state ids with the CSF for defined number of links
}
\description{
Determine class-stability factor (CSF) for specific number of links
}
\examples{
csf_n(addstates(wldf()),links=6)
}
