% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_isolates.R
\name{delete_isolates}
\alias{delete_isolates}
\title{Delete isolated nodes from network}
\usage{
delete_isolates(q)
}
\arguments{
\item{q}{an igraph object}
}
\value{
an igraph object containing only connected nodes
}
\description{
Delete isolated nodes from network
}
\examples{
g=igraph::graph.edgelist(cbind(a=c(1,2,3,4), b=c(3,1,1,4)),directed=TRUE)
delete_isolates(g)
}

