% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{di}
\alias{di}
\title{Calculate Dominance Index over time}
\usage{
di(df, plot = FALSE, type = "ds")
}
\arguments{
\item{df}{a dataframe containing three columns
the first column being the time, the second column the winner
and the second column being the loser}

\item{plot}{If equals TRUE will plot dominance over time}

\item{type}{If \strong{\code{lc}} will calculate the win proportion
dominance index used by Lindquist & Chase (2009). If \strong{\code{ds}}
will use David's Scores as the dominance index.}
}
\value{
a dataframe with the dominance index - a numeric value
between 0 and 1 for each animal at each timepoint
}
\description{
Calculate Dominance Index over time
}
\examples{
di(chickens[,-2], type="ds")
di(chickens[,-2], type="lc", plot=TRUE)
}
