% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_diffs.R
\name{edge_diffs}
\alias{edge_diffs}
\title{Determine edge differences between two igraph objects}
\usage{
edge_diffs(g1, g2)
}
\arguments{
\item{g1}{igraph object 1 - must be a subgraph of \strong{\code{g2}}}

\item{g2}{igraph object 2}
}
\value{
a vector of the number of same, different and new edges
}
\description{
Determine edge differences between two igraph objects
}
\examples{
g1 <- igraph::graph_from_literal( A -+ B -+ C)
    g2 <- igraph::graph_from_literal( A -+ B -+ C -+ A)
    edge_diffs(g1,g2)
}
