% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Example dataset of fish behavior}
\format{A data frame with 3189 observations and 4 variables:
\describe{
  \item{time}{Time in seconds of behavior}
  \item{winner}{Winner}
  \item{loser}{Loser}
  \item{behav}{Behavior: T = threat; B = bite; C = chase; L = lip-lock;}
}}
\usage{
fish
}
\description{
Dataframe with dominance behaviors directed from an actor
towards a recipient and timestampped.
}
\keyword{datasets}
