% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statespace.R
\name{get_statespace}
\alias{get_statespace}
\title{Get statespace characteristics of a directed graph}
\usage{
get_statespace(g)
}
\arguments{
\item{g}{a unweighted directed igraph object.}
}
\value{
a single row dataframe with characteristics of the graph
}
\description{
Get statespace characteristics of a directed graph
}
\examples{
g=igraph::graph.edgelist(cbind(a=c(1,2,3,3), b=c(4,1,1,2)),directed=TRUE)
get_statespace(g)
}

