% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interactions}
\alias{interactions}
\title{Identify which rows of win-loss dataframe are new or changed interactions}
\usage{
interactions(df)
}
\arguments{
\item{df}{a dataframe with two variables, winners and losers}
}
\value{
a dataframe with extra variables showing interaction number
    and if interaction is 'same', 'new' or a 'flip'.
}
\description{
Identify which rows of win-loss dataframe are new or changed interactions
}
\examples{
interactions(chickens[,3:4])
}
