% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions_diff.R
\name{interactions_diff}
\alias{interactions_diff}
\title{Identify which rows of win-loss dataframe are new or changed interactions}
\usage{
interactions_diff(df)
}
\arguments{
\item{df}{a dataframe with two variables, the first being winners and
the second being losers}
}
\value{
a vector with indices of rows that are new or changed interactions
}
\description{
Identify which rows of win-loss dataframe are new or changed interactions
}
\examples{
interactions_diff(chickens[,3:4])
}
