% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladybugs.R
\docType{data}
\name{ladybugs}
\alias{ladybugs}
\title{Example dataset of ladybug behavior}
\format{A data frame with 17 observations and 3 variables:
\describe{
  \item{time}{Time in seconds of behavior}
  \item{winner}{Winner}
  \item{loser}{Loser}
}}
\usage{
ladybugs
}
\description{
Dataframe with dominance behaviors directed from an winner
towards a loser and timestampped.
}
\keyword{datasets}
