% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastints.R
\name{lastints}
\alias{lastints}
\title{Get last interaction between each pair of nodes}
\usage{
lastints(df)
}
\arguments{
\item{df}{a dataframe with two variables, winners and losers}
}
\value{
a list with each element being the last interaction for each pair of nodes
up to that row of the dataframe
}
\description{
Get last interaction between each pair of nodes
}
\examples{
lastints(ladybugs[,2:3])
}
