% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastintsN.R
\name{lastintsN}
\alias{lastintsN}
\title{Get last N interactions between each pair of nodes}
\usage{
lastintsN(df, N = 3)
}
\arguments{
\item{df}{a dataframe with two variables, winners and losers}

\item{N}{the number of last interactions to get}
}
\value{
a list with each element being the last N interactions for each pair of nodes
up to that row of the dataframe
}
\description{
Get last N interactions between each pair of nodes
}
\examples{
lastintsN(ladybugs[,2:3],N=3)
}
