% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastnetN.R
\name{lastnetN}
\alias{lastnetN}
\title{Get network for last N interactions between each pair of nodes}
\usage{
lastnetN(df, N = 3, all = FALSE)
}
\arguments{
\item{df}{a dataframe with two variables, winners and losers}

\item{N}{the number of last interactions to get}

\item{all}{If = \strong{\code{FALSE}} then generates network based on one
directed edge with clear winner being individual with highest wins from
last N interactions between each pair. If = \strong{\code{TRUE}}
then keeps uses all edges from last N interactions between each pair.}
}
\value{
a list with each element being the network object derived from the last
    N interactions for each pair of nodes up to that row of the dataframe
}
\description{
Get network for last N interactions between each pair of nodes
}
\examples{
lastnetN(ladybugs[,2:3],N=3)
}
