% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn1.R
\name{nn1}
\alias{nn1}
\title{Get N/N(1) ratio}
\usage{
nn1(x)
}
\arguments{
\item{x}{a vector of integers}
}
\value{
the value of each integer divided by the number of integers equal to 1
}
\description{
Get N/N(1) ratio
}
\examples{
nn1(c(10,6,3,2,2,2,1,1,1,1,1,1))
}
