% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmatch.R
\name{rowmatch}
\alias{rowmatch}
\title{Match rows of dataframes}
\usage{
rowmatch(v1, v2, nomatch = NA)
}
\arguments{
\item{v1}{a dataframe or matrix}

\item{v2}{a row of a dataframe or matrix or a vector of length equal to ncol(v1)}

\item{nomatch}{how to represent non matching indices of rows}
}
\value{
a vector with a 1 in the index of matched rows
}
\description{
Match rows of dataframes
}
\examples{
v1 <- data.frame(A=c(1,2,3),B=c(2,3,3),C=c(3,3,2),D=c(1,1,1))
v2 <- data.frame(A=3,B=3,C=2,D=1)
rowmatch(v1,v2,nomatch=NA)
}
