% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_changes.R
\name{state_changes}
\alias{state_changes}
\title{Identify states that have changed from previous interaction}
\usage{
state_changes(df, subset = TRUE)
}
\arguments{
\item{df}{a dataframe produced from the \strong{\code{addstates()}} function}

\item{subset}{If \strong{\code{TRUE}} subset out rows where state change occurs
If \strong{\code{FALSE}} add column to original dataframe with state changes.}
}
\value{
either a subsetted dataframe with rows when states changed or
    an additional column to input dataframe stating if state change occurs
}
\description{
Identify states that have changed from previous interaction
}
\examples{
state_changes(addstates(ladybugs[,2:3]))
state_changes(addstates(ladybugs[,2:3]),subset=FALSE)
}
