% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_plot.R
\name{state_plot}
\alias{state_plot}
\title{Create state space plot for specific interaction}
\usage{
state_plot(df, N = 1)
}
\arguments{
\item{df}{A dataframe of two variables - winner and loser.
Individuals should be named 1-4.}

\item{N}{Row of the dataframe to plot the state space for.}
}
\value{
a plot of the state space at row If \strong{\code{N}} of the dataframe
}
\description{
Create state space plot for specific interaction
}
\examples{
state_plot(fish[1:20,2:3], N=20)
}
