% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_plot_grid.R
\name{state_plot_grid}
\alias{state_plot_grid}
\title{Create state space plot grid for specific interactions}
\usage{
state_plot_grid(df, nc = NULL, nr = NULL, textsize = 10, labelno = 1)
}
\arguments{
\item{df}{A dataframe of tWO variables - winner and loser.
Individuals should be named 1-4.}

\item{nc}{Number of columns in the grid}

\item{nr}{Number of rows in the grid}

\item{textsize}{Font size of labels}

\item{labelno}{Label to start plot at}
}
\value{
a plot of the state space at row If \strong{\code{N}} of the dataframe
}
\description{
Create state space plot grid for specific interactions
}
\examples{
state_plot_grid(fish[1:30,2:3], nc=6, nr=5)
}
