% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateclass.R
\name{stateclass}
\alias{stateclass}
\title{Find State Class}
\usage{
stateclass(x)
}
\arguments{
\item{x}{a vector of network characteristics generated by get_statespace function}
}
\value{
the state class corresponding to the network characteristics
}
\description{
Find State Class
}
\examples{
g=igraph::graph.edgelist(cbind(a=c(1,2,3,3), b=c(4,1,1,2)),directed=TRUE)
gs=netchar(g)
stateclass(gs)
}

