% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateclasses.R
\docType{data}
\name{stateclasses}
\alias{stateclasses}
\title{State classes of 41 network states}
\format{A vector of length 41:}
\usage{
stateclasses
}
\description{
Vector defining the classes of 41 network states
}
\keyword{datasets}
