% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateid.R
\name{stateid}
\alias{stateid}
\title{Find State ID}
\usage{
stateid(x)
}
\arguments{
\item{x}{a vector of network characteristics generated by get_statespace function}
}
\value{
the state id corresponding to the network characteristics
}
\description{
Find State ID
}
\examples{
g=igraph::graph.edgelist(cbind(a=c(1,2,3,3), b=c(4,1,1,2)),directed=TRUE)
gs=get_statespace(g)
stateid(gs)
}

