% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad_duration.R
\name{triad_duration}
\alias{triad_duration}
\title{Calculate Dominance Index over time}
\usage{
triad_duration(df)
}
\arguments{
\item{df}{a dataframe containing two columns -
the first column being the winner
and the second column being the loser}
}
\value{
a list of dataframes with each element being
    individual triads. Each dataframe describes the type of
    edge (new,same,flip), the state of the triad (single,
    double dominant, double subordinate, pass-along, transitive,
    cycle), and the number of nodes and edges in the triad at
    each row.
}
\description{
Calculate Dominance Index over time
}
\examples{
triad_duration(chickens[,3:4])
}
