% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wldf.R
\name{wldf}
\alias{wldf}
\title{Generate random winner-loser dataframe}
\usage{
wldf(ints = 100, actors = 4, intprob = NULL)
}
\arguments{
\item{ints}{number of interactions}

\item{actors}{number of individuals}

\item{intprob}{vector of interaction probabilities of same length as `actors`}
}
\value{
a dataframe with two variables, winner and loser
}
\description{
Generate random winner-loser dataframe
}
\examples{
df<-data.frame(winner=c(1,2,3,1,2,3,2,1,2,3,3,1,2,3,4,3,1,3,2,1,1,1,1,2,2),
loser=c(2,4,4,4,3,2,1,3,3,4,4,2,3,2,3,4,2,4,3,3,3,2,2,4,3)
)
wldf()
wldf(ints=30, actors=3, intprob=c(0.8,0.1,0.1))
}

